Externalize class
