+++ wsmaster/che-core-api-factory-github/src/main/java/org/eclipse/che/api/factory/server/github/GithubFactoryParametersResolver.java
@@ -15,9 +15,6 @@ import static org.eclipse.che.api.factory.shared.Constants.CURRENT_VERSION;
 import static org.eclipse.che.api.factory.shared.Constants.URL_PARAMETER_NAME;
 import static org.eclipse.che.dto.server.DtoFactory.newDto;
 
-import java.io.IOException;
-import java.net.URI;
-import java.net.URISyntaxException;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
@@ -30,9 +27,7 @@ import org.eclipse.che.api.factory.server.DefaultFactoryParameterResolver;
 import org.eclipse.che.api.factory.server.urlfactory.ProjectConfigDtoMerger;
 import org.eclipse.che.api.factory.server.urlfactory.URLFactoryBuilder;
 import org.eclipse.che.api.factory.shared.dto.FactoryDto;
-import org.eclipse.che.api.workspace.server.devfile.FileContentProvider;
 import org.eclipse.che.api.workspace.server.devfile.URLFetcher;
-import org.eclipse.che.api.workspace.server.devfile.exception.DevfileException;
 import org.eclipse.che.api.workspace.shared.dto.ProjectConfigDto;
 import org.eclipse.che.api.workspace.shared.dto.devfile.ProjectDto;
 
@@ -98,7 +93,7 @@ public class GithubFactoryParametersResolver extends DefaultFactoryParameterReso
         urlFactoryBuilder
             .createFactoryFromDevfile(
                 githubUrl,
-                new GithubFileContentProvider(githubUrl),
+                new GithubFileContentProvider(githubUrl, urlFetcher),
                 extractOverrideParams(factoryParameters))
             .orElseGet(() -> newDto(FactoryDto.class).withV(CURRENT_VERSION).withSource("repo"));
 
@@ -139,32 +134,5 @@ public class GithubFactoryParametersResolver extends DefaultFactoryParameterReso
           });
     }
     return factory;
-  }
-
-  /** Github specific file content provider. */
-  private class GithubFileContentProvider implements FileContentProvider {
-
-    private final GithubUrl githubUrl;
-
-    private GithubFileContentProvider(GithubUrl githubUrl) {
-      this.githubUrl = githubUrl;
-    }
-
-    @Override
-    public String fetchContent(String fileURL) throws IOException, DevfileException {
-      {
-        String requestURL;
-        try {
-          if (new URI(fileURL).isAbsolute()) {
-            requestURL = fileURL;
-          } else {
-            requestURL = githubUrl.rawFileLocation(fileURL);
-          }
-        } catch (URISyntaxException e) {
-          throw new DevfileException(e.getMessage(), e);
-        }
-        return urlFetcher.fetch(requestURL);
-      }
-    }
   }
 }
